+++ examples/examples_modeling_tutorial/TestRun_NumBeam.py
@@ -139,6 +139,8 @@ primary =  CustomHotRegion(bounds=bounds,
 	                    num_leaves=100,
 	                    num_rays=200,                            
                             fbeam=True,
+                            atm_ext="Num4D",
+                            beam_opt=2,
                             prefix='p')
               	                    
 bounds['super_abb'] = None  
@@ -195,6 +197,8 @@ secondary =  CustomHotRegion(bounds=bounds,
 	                    do_fast=False,
 	                    is_antiphased=True,
                             fbeam=True,
+                            atm_ext="Num4D",
+                            beam_opt=2,
                             prefix='s')